refactor: actualización completa de controladores REST
 Mejoras técnicas:
- Parámetros @PathVariable/@RequestParam: Integer  Long
- Tipos de fecha modernos: java.sql.Date  LocalDate, String  LocalTime
- Métodos de entidad actualizados: getPregunta()  getPreguntaSeguridad()
- Imports optimizados: eliminación de Calendar, Date deprecated
- Validación mejorada: String.isEmpty() en lugar de != ''

 API REST optimizada:
- Endpoints más robustos con tipos Long escalables
- Manejo temporal preciso con LocalDate/LocalTime
- Responses consistentes con entidades refactorizadas
- Mejor integración con frontend React SPA

 Controladores afectados:
- AuthController: métodos de seguridad corregidos
- CitaOdontologicaController: gestión de citas optimizada
- ConsultorioController: agenda por fechas mejorada
- Todos los CRUD controllers: tipos Long consistentes

+++ src/main/java/com/example/demo/controller/MedicamentoController.java
@@ -35,7 +35,7 @@ public class MedicamentoController {
 	}
 
 	@DeleteMapping(path = "/delete")
-	public @ResponseBody String eliminar(@RequestParam Integer id) {
+	public @ResponseBody String eliminar(@RequestParam Long id) {
 		if (medicamentoRepository.existsById(id)) {
 			medicamentoRepository.deleteById(id);
 			return "Medicamento eliminado exitosamente";
